code style clean up

+++ app/src/main/java/com/hover/stax/channels/ChannelViewModel.java
@@ -29,9 +29,13 @@ public class ChannelViewModel extends AndroidViewModel {
 		loadSims(application);
 	}
 
-	LiveData<List<Channel>> getChannels() { return channels; }
+	LiveData<List<Channel>> getChannels() {
+		return channels;
+	}
 
-	public MutableLiveData<List<SimInfo>> getSims() { return sims; }
+	public MutableLiveData<List<SimInfo>> getSims() {
+		return sims;
+	}
 
 	private void loadChannels() {
 		if (channels == null) {
@@ -40,15 +44,15 @@ public class ChannelViewModel extends AndroidViewModel {
 		channels = repo.getAll();
 		getPendingSelected();
 	}
+
 	void loadInitiallySelectedChannels(List<Channel> channels) {
 		List<Integer> ls = new ArrayList<>();
-		for(Channel channel: channels) {
-			if(channel.selected) ls.add(channel.id);
+		for (Channel channel : channels) {
+			if (channel.selected) ls.add(channel.id);
 		}
 		selected.postValue(ls);
 
 	}
-
 
 
 	public LiveData<List<Integer>> getPendingSelected() {
@@ -76,8 +80,8 @@ public class ChannelViewModel extends AndroidViewModel {
 
 	public void saveSelected() {
 		List<Channel> allChannels = channels.getValue() != null ? channels.getValue() : new ArrayList<>();
-		for (Channel channel: allChannels) {
-			if (selected.getValue().contains((Integer) channel.id)) {
+		for (Channel channel : allChannels) {
+			if (selected.getValue().contains(channel.id)) {
 				channel.selected = true;
 				repo.update(channel);
 			}
